Parser : map xsd:decimal to Currency.
git-svn-id: http://svn.code.sf.net/p/lazarus-ccr/svn@3052 8e941d3f-bd1b-0410-a28a-d453659cc2b4

+++ wst/trunk/ws_helper/pascal_parser_intf.pas
@@ -277,7 +277,7 @@ const
           ('Qword', 'TComplexInt64UContentRemotable', 'unsignedLong'),
           ('Single', 'TComplexFloatSingleContentRemotable', 'float'),
           ('Double', 'TComplexFloatDoubleContentRemotable', 'double'),
-          ('Extended', 'TComplexFloatExtendedContentRemotable', 'decimal')
+          ('Extended', 'TComplexFloatExtendedContentRemotable', 'double')
         );
    BOXED_TYPES_COUNT = 2;
      BOXED_TYPES : Array[0..Pred(BOXED_TYPES_COUNT)] Of array[0..2] of string = (
